feat(game): ajouter la méthode d'édition de game pour la BD

+++ client/src/app/pages/game-creation-page/game-creation-page.component.ts
@@ -1,4 +1,5 @@
 import { Component, OnInit } from '@angular/core';
+import { RouterLink } from '@angular/router';
 import { GameCardComponent } from '@app/components/game/game-card/game-card.component';
 import { GameService } from '@app/services/game.service';
 import { Game } from '@common/classes/game';
@@ -6,7 +7,7 @@ import { Game } from '@common/classes/game';
 
 @Component({
   selector: 'app-game-creation-page',
-  imports: [GameCardComponent],
+  imports: [GameCardComponent, RouterLink],
   templateUrl: './game-creation-page.component.html',
   styleUrl: './game-creation-page.component.scss',
 })
@@ -15,7 +16,6 @@ export class GameCreationPageComponent implements OnInit {
   constructor(private gameService: GameService) {}
 
   games: Game[] = [];
-  game: Game;
 
   ngOnInit(): void {
     this.getAllGames();
@@ -28,21 +28,37 @@ export class GameCreationPageComponent implements OnInit {
   }
 
   addGame(game: Game): void {
+    //game deja présente
+    if (this.games.some(x => x._id === game._id)) return;
+
     this.gameService.addGame(game).subscribe(() => {
       this.getAllGames();
     });
   }
 
   getGame(_id: string): void {
+    //si deja présente
+    if (this.games.some(x => x._id === _id)) return;
     this.gameService.getGame(_id).subscribe(oneGame => {
       this.games.push(oneGame);
     });
   }
 
   deleteGame(_id: string): void {
+    //game pas présente
+    const y = this.games.find((x) => x._id === _id);
+    if (y && !this.games.some(x => x._id === y._id)) return;
+
     this.gameService.deleteGame(_id).subscribe(() => {
       this.getAllGames();
     });
-
   }
+
+
+  changeGameVisibility(_id: string, visibility: boolean): void {
+    this.gameService.patchGame(_id, visibility).subscribe(() => {
+      this.getAllGames();
+    });
+  }
+
 }
